requirements.txt
openai
PyPDF2
chromadb
python-dotenv